pyproject.toml
[project]
name = "analytics-engineering-modern-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
    "dbt-core>=1.10.11",
    "dbt-duckdb>=1.9.6",
    "faker>=37.6.0",
    "numpy==2.3.3",
    "pandas==2.3.2",
    "pyarrow>=21.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
[tool.ruff]
line-length = 120

